day23/pyproject.toml
[project]
name = "day23"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.27.1",
    "litellm>=1.82.3",
    "openai>=2.23.0",
]
